feat: Introduce new environment variable
This commit introduces the new env variable CONGRESS_API_KEY, needed for the upcoming module that will get congress data, like members, bills, etc from the congress api.

The "requests" library is added as well to the requirements.txt file and more file extensions are added to the .gitignore file.