refactor: Replace INSERT-IGNORE queries with UPSERT queries
This commit replaces the INSERT-IGNORE queries in all the loading functions in db_ops.py with UPSERT (INSERT/ON CONFLICT/UPDATE) queries since these are more robust and don't allows error to pass silently during query execution.

Additionally a new env variable is added: SETUP_SCHEMA. This variable defines if the database has to be initialized from scratch when the main process is executed.